Added a middleware to log any request errors so those can be shown.
Also added a QWCMessage model that we will use to log all the messages
we received from the Web Connector. Idea is that we'll have a nice
description of each so that between both, we can provide some useful
feedback as to what's happening between the portal and the web
connector.